fixed head conflicts
